requirements.txt
flask
flask-login
pyodbc
python-decouple
openpyxl
pandas
werkzeug
